+++ api/migrations/1601148672971-initialSchema.ts
@@ -1,7 +1,7 @@
 import {MigrationInterface, QueryRunner} from "typeorm";
 
-export class initialSchema1601148592825 implements MigrationInterface {
-    name = 'initialSchema1601148592825'
+export class initialSchema1601148672971 implements MigrationInterface {
+    name = 'initialSchema1601148672971'
 
     public async up(queryRunner: QueryRunner): Promise<void> {
         await queryRunner.query(`CREATE TYPE "user_roles_enum" AS ENUM('ADMIN', 'MODERATOR', 'USER')`);
@@ -11,7 +11,8 @@ export class initialSchema1601148592825 implements MigrationInterface {
         await queryRunner.query(`CREATE TYPE "reaction_type_enum" AS ENUM('APPROVE', 'REFUTE', 'SKEPTIC')`);
         await queryRunner.query(`CREATE TABLE "reaction" ("id" SERIAL NOT NULL, "type" "reaction_type_enum", "created" TIMESTAMP NOT NULL DEFAULT now(), "user_id" integer NOT NULL, "comment_id" integer NOT NULL, CONSTRAINT "UQ_19601094466e474ac560d4fc9db" UNIQUE ("user_id", "comment_id"), CONSTRAINT "PK_41fbb346da22da4df129f14b11e" PRIMARY KEY ("id"))`);
         await queryRunner.query(`CREATE TABLE "comment" ("id" SERIAL NOT NULL, "created" TIMESTAMP NOT NULL DEFAULT now(), "updated" TIMESTAMP NOT NULL DEFAULT now(), "score" integer NOT NULL DEFAULT 0, "author_id" integer NOT NULL, "comments_area_id" integer NOT NULL, "message_id" integer, "parent_id" integer, CONSTRAINT "REL_a982661c7375f3b20e0eb9ed19" UNIQUE ("message_id"), CONSTRAINT "PK_0b0e4bbc8415ec426f87f3a88e2" PRIMARY KEY ("id"))`);
-        await queryRunner.query(`CREATE TABLE "report" ("id" SERIAL NOT NULL, "message" text, "waitingForReview" boolean NOT NULL, "created" TIMESTAMP NOT NULL DEFAULT now(), "user_id" integer NOT NULL, "comment_id" integer NOT NULL, CONSTRAINT "UQ_27abeaa95cd2bb61c1125b173ee" UNIQUE ("user_id", "comment_id"), CONSTRAINT "PK_99e4d0bea58cba73c57f935a546" PRIMARY KEY ("id"))`);
+        await queryRunner.query(`CREATE TYPE "report_moderationaction_enum" AS ENUM('IGNORE')`);
+        await queryRunner.query(`CREATE TABLE "report" ("id" SERIAL NOT NULL, "message" text, "moderationAction" "report_moderationaction_enum", "created" TIMESTAMP NOT NULL DEFAULT now(), "reporter_id" integer, "comment_id" integer, "moderator_id" integer, CONSTRAINT "UQ_b6a6103bbd3ebdfd0a2d0eaa2f4" UNIQUE ("reporter_id", "comment_id"), CONSTRAINT "PK_99e4d0bea58cba73c57f935a546" PRIMARY KEY ("id"))`);
         await queryRunner.query(`CREATE TABLE "subscription" ("id" SERIAL NOT NULL, "created" TIMESTAMP NOT NULL DEFAULT now(), "user_id" integer NOT NULL, "comment_id" integer, CONSTRAINT "UQ_42aee501742b3fad49552712551" UNIQUE ("user_id", "comment_id"), CONSTRAINT "PK_8c3e00ebd02103caa1174cd5d9d" PRIMARY KEY ("id"))`);
         await queryRunner.query(`CREATE TYPE "open_comments_area_request_status_enum" AS ENUM('PENDING', 'APPROVED', 'REFUSED')`);
         await queryRunner.query(`CREATE TABLE "open_comments_area_request" ("id" SERIAL NOT NULL, "identifier" character varying NOT NULL, "created" TIMESTAMP NOT NULL DEFAULT now(), "updated" TIMESTAMP NOT NULL DEFAULT now(), "status" "open_comments_area_request_status_enum" NOT NULL, "requester_id" integer, CONSTRAINT "PK_e16852322c928ef40b4d5532be1" PRIMARY KEY ("id"))`);
@@ -26,8 +27,9 @@ export class initialSchema1601148592825 implements MigrationInterface {
         await queryRunner.query(`ALTER TABLE "comment" ADD CONSTRAINT "FK_e6f3a528893a48ee11a1d8d5d77" FOREIGN KEY ("comments_area_id") REFERENCES "comments_area"("id") ON DELETE NO ACTION ON UPDATE NO ACTION`);
         await queryRunner.query(`ALTER TABLE "comment" ADD CONSTRAINT "FK_a982661c7375f3b20e0eb9ed19e" FOREIGN KEY ("message_id") REFERENCES "message"("id") ON DELETE NO ACTION ON UPDATE NO ACTION`);
         await queryRunner.query(`ALTER TABLE "comment" ADD CONSTRAINT "FK_8bd8d0985c0d077c8129fb4a209" FOREIGN KEY ("parent_id") REFERENCES "comment"("id") ON DELETE NO ACTION ON UPDATE NO ACTION`);
-        await queryRunner.query(`ALTER TABLE "report" ADD CONSTRAINT "FK_c6686efa4cd49fa9a429f01bac8" FOREIGN KEY ("user_id") REFERENCES "user"("id") ON DELETE NO ACTION ON UPDATE NO ACTION`);
+        await queryRunner.query(`ALTER TABLE "report" ADD CONSTRAINT "FK_d41df66b60944992386ed47cf2e" FOREIGN KEY ("reporter_id") REFERENCES "user"("id") ON DELETE NO ACTION ON UPDATE NO ACTION`);
         await queryRunner.query(`ALTER TABLE "report" ADD CONSTRAINT "FK_8bb2bc4a3d9c55e031bc5d015c5" FOREIGN KEY ("comment_id") REFERENCES "comment"("id") ON DELETE NO ACTION ON UPDATE NO ACTION`);
+        await queryRunner.query(`ALTER TABLE "report" ADD CONSTRAINT "FK_10a121780b323901bc3ebc7fa69" FOREIGN KEY ("moderator_id") REFERENCES "user"("id") ON DELETE NO ACTION ON UPDATE NO ACTION`);
         await queryRunner.query(`ALTER TABLE "subscription" ADD CONSTRAINT "FK_940d49a105d50bbd616be540013" FOREIGN KEY ("user_id") REFERENCES "user"("id") ON DELETE NO ACTION ON UPDATE NO ACTION`);
         await queryRunner.query(`ALTER TABLE "subscription" ADD CONSTRAINT "FK_3292074e41c7fe52ab5374367db" FOREIGN KEY ("comment_id") REFERENCES "comment"("id") ON DELETE NO ACTION ON UPDATE NO ACTION`);
         await queryRunner.query(`ALTER TABLE "open_comments_area_request" ADD CONSTRAINT "FK_9c04cd14d9ea54b09dc4e89d717" FOREIGN KEY ("requester_id") REFERENCES "user"("id") ON DELETE NO ACTION ON UPDATE NO ACTION`);
@@ -39,8 +41,9 @@ export class initialSchema1601148592825 implements MigrationInterface {
         await queryRunner.query(`ALTER TABLE "open_comments_area_request" DROP CONSTRAINT "FK_9c04cd14d9ea54b09dc4e89d717"`);
         await queryRunner.query(`ALTER TABLE "subscription" DROP CONSTRAINT "FK_3292074e41c7fe52ab5374367db"`);
         await queryRunner.query(`ALTER TABLE "subscription" DROP CONSTRAINT "FK_940d49a105d50bbd616be540013"`);
+        await queryRunner.query(`ALTER TABLE "report" DROP CONSTRAINT "FK_10a121780b323901bc3ebc7fa69"`);
         await queryRunner.query(`ALTER TABLE "report" DROP CONSTRAINT "FK_8bb2bc4a3d9c55e031bc5d015c5"`);
-        await queryRunner.query(`ALTER TABLE "report" DROP CONSTRAINT "FK_c6686efa4cd49fa9a429f01bac8"`);
+        await queryRunner.query(`ALTER TABLE "report" DROP CONSTRAINT "FK_d41df66b60944992386ed47cf2e"`);
         await queryRunner.query(`ALTER TABLE "comment" DROP CONSTRAINT "FK_8bd8d0985c0d077c8129fb4a209"`);
         await queryRunner.query(`ALTER TABLE "comment" DROP CONSTRAINT "FK_a982661c7375f3b20e0eb9ed19e"`);
         await queryRunner.query(`ALTER TABLE "comment" DROP CONSTRAINT "FK_e6f3a528893a48ee11a1d8d5d77"`);
@@ -56,6 +59,7 @@ export class initialSchema1601148592825 implements MigrationInterface {
         await queryRunner.query(`DROP TYPE "open_comments_area_request_status_enum"`);
         await queryRunner.query(`DROP TABLE "subscription"`);
         await queryRunner.query(`DROP TABLE "report"`);
+        await queryRunner.query(`DROP TYPE "report_moderationaction_enum"`);
         await queryRunner.query(`DROP TABLE "comment"`);
         await queryRunner.query(`DROP TABLE "reaction"`);
         await queryRunner.query(`DROP TYPE "reaction_type_enum"`);