+++ api/src/modules/comment/report/report.entity.ts
@@ -3,26 +3,34 @@ import { Column, CreateDateColumn, Entity, JoinColumn, ManyToOne, PrimaryGenerat
 import { User } from '../../user/user.entity';
 import { Comment } from '../comment.entity';
 
+export enum ReportModerationAction {
+  IGNORED = 'IGNORE',
+}
+
 @Entity({ name: 'report' })
-@Unique(['user', 'comment'])
+@Unique(['reportedBy', 'comment'])
 export class Report {
 
   @PrimaryGeneratedColumn()
   id: number;
 
-  @ManyToOne(type => User, { nullable: false, eager: true })
-  @JoinColumn({ name: 'user_id' })
-  user: User;
+  @ManyToOne(type => User, { eager: true })
+  @JoinColumn({ name: 'reporter_id' })
+  reportedBy: User;
 
-  @ManyToOne(type => Comment, { nullable: false, eager: true })
+  @ManyToOne(type => Comment, { eager: true })
   @JoinColumn({ name: 'comment_id' })
   comment: Comment;
 
   @Column({ type: 'text', nullable: true })
   message: string;
 
-  @Column()
-  waitingForReview: boolean;
+  @ManyToOne(type => User, { nullable: true, eager: true })
+  @JoinColumn({ name: 'moderator_id' })
+  moderatedBy: User;
+
+  @Column({ type: 'enum', enum: ReportModerationAction, nullable: true })
+  moderationAction: ReportModerationAction;
 
   @CreateDateColumn()
   created: Date;