+++ api/src/modules/comment/report/report.module.ts
@@ -1,7 +1,5 @@
-import { forwardRef, Module } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-
-import { CommentModule } from '../comment.module';
 
 import { Report } from './report.entity';
 import { ReportFactory } from './report.factory';
@@ -10,7 +8,6 @@ import { ReportService } from './report.service';
 @Module({
   imports: [
     TypeOrmModule.forFeature([Report]),
-    forwardRef(() => CommentModule),
   ],
   providers: [
     ReportService,