+++ api/src/modules/moderation/dtos/ignore-reports-in.dto.ts
@@ -1,0 +1,8 @@
+import { IsInt } from 'class-validator';
+
+export class IgnoreReportsInDto {
+
+  @IsInt()
+  commentId: number;
+
+}