+++ app/src/pages/Moderation/ReportedComments/index.tsx
@@ -3,13 +3,19 @@ import React from 'react';
 import AsyncContent from '../../../components/AsyncContent';
 import Fallback from '../../../components/Fallback';
 import useAxiosPaginated from '../../../hooks/use-axios-paginated';
-import { parseReportedComment } from '../../../types/Report';
+import useEditableDataset from '../../../hooks/use-editable-dataset';
+import { parseReportedComment, ReportedComment as ReportedCommentType } from '../../../types/Report';
 import Section from '../Section';
 
 import ReportedComment from './ReportedComment';
 
 const ReportedComments: React.FC = () => {
-  const [{ data: reportedComments, loading }] = useAxiosPaginated('/api/moderation/reports', parseReportedComment);
+  const [{ data, loading }] = useAxiosPaginated('/api/moderation/reports', parseReportedComment);
+  const [reportedComments, { remove }] = useEditableDataset(data);
+
+  const handleModerated = (comment: ReportedCommentType) => {
+    remove(comment);
+  };
 
   return (
     <Section title="Commentaires signalés">
@@ -22,8 +28,13 @@ const ReportedComments: React.FC = () => {
           >
             {() => (
               <>
-                {reportedComments.map(({ reports, ...comment }) => (
-                  <ReportedComment key={comment.id} comment={comment} reports={reports} />
+                {reportedComments.map((comment) => (
+                  <ReportedComment
+                    key={comment.id}
+                    comment={comment}
+                    reports={comment.reports}
+                    onModerated={() => handleModerated(comment)}
+                  />
                 ))}
               </>
             )}