+++ app/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - List reported comments
+- Ignore reported comments
 
 ## [0.5.0] - 2020-09-22